+++ tests/unit/igraph_pagerank.c
@@ -105,9 +105,10 @@ int main() {
     printf("PRPACK: "); print_vector(&res);
     IGRAPH_ASSERT(is_almost_one(value));
 
-    /* Check twice more for consistency */
+    /* Check twice more for consistency, this time without explicitly
+     * supplied ARPACK options */
     igraph_pagerank(&g, IGRAPH_PAGERANK_ALGO_ARPACK, &res, &value,
-                    igraph_vss_all(), 0, 0.85, 0, &arpack_options);
+                    igraph_vss_all(), 0, 0.85, 0, 0);
     printf("ARPACK: "); print_vector(&res);
     IGRAPH_ASSERT(is_almost_one(value));
 
@@ -117,7 +118,7 @@ int main() {
     IGRAPH_ASSERT(is_almost_one(value));
 
     igraph_pagerank(&g, IGRAPH_PAGERANK_ALGO_ARPACK, &res, &value,
-                    igraph_vss_all(), 0, 0.85, 0, &arpack_options);
+                    igraph_vss_all(), 0, 0.85, 0, 0);
     printf("ARPACK: "); print_vector(&res);
     IGRAPH_ASSERT(is_almost_one(value));